Restructure project, add enrollment notification
