+++ src/components/enrollment-form/success-screen/animations/checkmark.json
