+++ src/components/page/footer/index.tsx
